# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for e-learning platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments:
    post:
      operationId: submitPayment
      tags: [payments]
      summary: Submit payment with screenshot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitPaymentRequest"
      responses:
        "201":
          description: Payment submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
...
  /payments/{id}/validate:
    post:
      operationId: validatePayment
      tags: [payments]
      summary: Validate a payment (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment validated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /payments/{id}/reject:
    post:
      operationId: rejectPayment
      tags: [payments]
      summary: Reject a payment (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /upload/payment-screenshot:
    post:
      operationId: uploadPaymentScreenshot
      tags: [payments]
      summary: Upload payment screenshot
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "200":
          description: Upload successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    UploadResponse:
      type: object
      properties:
        url:
          type: string
      required:
        - url
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        typeAbonnement:
          type: string
          enum: [1_mois, 3_mois, 9_mois]
        montant:
          type: number
        date:
          type: string
          format: date-time
        statut:
          type: string
          enum: [en_attente, confirmé, rejeté]
        captureEcran:
          type: string
          nullable: true
      required:
        - id
        - userId
        - typeAbonnement
        - montant
        - date
        - statut
...
    SubmitPaymentRequest:
      type: object
      properties:
        typeAbonnement:
          type: string
          enum: [1_mois, 3_mois, 9_mois]
        captureEcran:
          type: string
          nullable: true
      required:
        - typeAbonnement